admin-user

+++ src/main/java/com/xx/idoctorall/dao/api/CommentRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CommentRepository extends CrudRepository<Comment, Long> {
+public interface CommentRepository extends CrudRepository<Comment, Integer> {
 
-    List<Comment> findByUserid(Long userid);
+    List<Comment> findByUserid(int userid);
     Comment save(Comment comment);
 }
